% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_workflow.R
\name{accuracy_metric}
\alias{accuracy_metric}
\title{Create Accuracy Metric Set}
\usage{
accuracy_metric(DF)
}
\arguments{
\item{DF}{The Predicted Data Frame}
}
\value{
The Accuracy metric set
}
\description{
Create Accuracy Metric Set
}
