% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{create_recipe}
\alias{create_recipe}
\title{Create Recipe for Model}
\usage{
create_recipe(DF)
}
\arguments{
\item{DF}{The Data Frame to create recipe}
}
\value{
Recipe object
}
\description{
This functions is a feature engineering recipe before model
}
