% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_workflow.R
\name{create_vip_dt}
\alias{create_vip_dt}
\title{The function to create Variable Importance Data}
\usage{
create_vip_dt(modeltime_table)
}
\arguments{
\item{modeltime_table}{The fitted modeltime workflowset}
}
\value{
The Data Frame of Variable Importance
}
\description{
The function to create Variable Importance Data
}
