% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepare.R
\name{data_prepare}
\alias{data_prepare}
\title{Initialize Data}
\usage{
data_prepare(data_list = as.list())
}
\arguments{
\item{data_list}{The raw data list from function \code{read_raw_data}}
}
\value{
Complete Data

Future Data with Features
}
\description{
Initialize Data
}
