% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_workflow.R
\name{forecast_values}
\alias{forecast_values}
\title{Forecast the values}
\usage{
forecast_values(model_fit, DF)
}
\arguments{
\item{model_fit}{The fitted model workflow sets is output of \code{model_fit}}

\item{DF}{The Testing / New Data Frame to predict.}
}
\value{
Predicted Data Frame
}
\description{
Forecast the values
}
