% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_forecast.R
\name{init_forecast}
\alias{init_forecast}
\title{Creates forecast values}
\usage{
init_forecast(DF = is.data.frame(), new_df = NULL, model_name = is.character())
}
\arguments{
\item{DF}{The training Data Frame}

\item{new_df}{The Data to predict future values}

\item{model_name}{The model names}
}
\value{
Fitted Modeltime Table

Test Data Frame's Forecast Values

Accuracy Table
}
\description{
This function join all the model workflow functions
}
