% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_prepare.R
\name{init_prepare}
\alias{init_prepare}
\title{Initialize Data}
\usage{
init_prepare(data_list = is.list())
}
\arguments{
\item{data_list}{The raw data list from function \code{read_raw_data}}
}
\value{
Complete Data

Future Data with Features
}
\description{
Initialize Data
}
