% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_raw_data.R
\name{init_raw_data}
\alias{init_raw_data}
\title{Read data}
\usage{
init_raw_data(
  train_raw_path = is.character(),
  test_raw_path = is.character(),
  meal_info_path = is.character()
)
}
\arguments{
\item{train_raw_path}{The location of train .csv file.}

\item{test_raw_path}{The location of test .csv file.}

\item{meal_info_path}{The location of meal master csv file.}
}
\value{
List of data.

train_data: The data.table of training data.

test_data: The data.table of test data.
}
\description{
Read raw Training & test data of food demand provided
by AnalyticsVidhya.
}
