% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_list.R
\name{ml_lgbm}
\alias{ml_lgbm}
\title{Tidymodel Models --------------------------------------------------------}
\usage{
ml_lgbm()
}
\value{
model
}
\description{
#' Stan Regression
#'
#' @return model
#' @export
ml_stan <- function(){
parsnip::linear_reg(mode = "regression") |>
parsnip::set_engine("stan")
}
}
\details{
#' Decision Tree
#'
#' @return model
#' @export
ml_dt <- function(){
parsnip::decision_tree(mode = "regression") |>
parsnip::set_engine("rpart")
}

LightGBM
}
