% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_workflow.R
\name{model_fit}
\alias{model_fit}
\title{Fit the workflow set}
\usage{
model_fit(model_set, DF)
}
\arguments{
\item{model_set}{The workflow set which is output of \code{model_workflow}}

\item{DF}{The training Data Frame}
}
\value{
Workflow Set with fit column
}
\description{
Fit the workflow set
}
