% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_workflow.R
\name{model_workflow}
\alias{model_workflow}
\title{Create workflow sets}
\usage{
model_workflow(DF = is.data.frame(), model_name = is.character())
}
\arguments{
\item{DF}{The training Data Frame}

\item{model_name}{The model names}
}
\value{
Workflow Set object
}
\description{
Create workflow sets
}
