% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_visuals.R
\name{plot_forecast}
\alias{plot_forecast}
\title{The function to plot time series against forecast vs actuals}
\usage{
plot_forecast(fcast_df, actual_df)
}
\arguments{
\item{fcast_df}{The Forecast Data Frame}

\item{actual_df}{The Actual Data Frame}
}
\value{
plotly object
}
\description{
The function to plot time series against forecast vs actuals
}
