% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_data.R
\name{read_raw_data}
\alias{read_raw_data}
\title{Read data}
\usage{
read_raw_data(
  train_raw_path = as.character(),
  test_raw_path = as.character(),
  meal_info_path = as.character()
)
}
\arguments{
\item{train_raw_path}{The location of train .csv file.}

\item{test_raw_path}{The location of test .csv file.}

\item{meal_info_path}{The location of meal master csv file.}
}
\value{


List of data.

train_data: The data.table of training data.

test_data: The data.table of test data.
}
\description{
Read raw Training & test data of food demand provided
by AnalyticsVidhya.
}
\examples{
data_list = read_raw_data(
     train_raw_path = "inst/extdata/train.csv",
     test_raw_path = "inst/extdata/test_QoiMO9B.csv",
     meal_info_path = "inst/extdata/meal_info.csv"
)
}
