% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_long_short.R
\name{separate_long_short}
\alias{separate_long_short}
\title{Separate Train data}
\usage{
separate_long_short(data_list = as.list())
}
\arguments{
\item{data_list}{The raw data list from function \code{read_raw_data}}
}
\value{


List of data list.

long_ts: The data with greater than 20 transactions.

short_ts: The data with shorter than 20 transactions.
}
\description{
Ingest raw train data and separate the data into long (> 20 transactions)
& short (<= 20 time series) time series data.
}
\examples{
\dontrun{
data_list = separate_long_short(data_list = train_data_list)
}
}
