% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{ts_split}
\alias{ts_split}
\title{Creates Time Series Split for 10 weeks}
\usage{
ts_split(DF)
}
\arguments{
\item{DF}{The Data Frame to split}
}
\value{
Train/Test split
}
\description{
Creates Time Series Split for 10 weeks
}
